mne_create_object extension
